follow lint warnings

+++ realm/realm-library/src/main/java/io/realm/internal/TableView.java
@@ -818,7 +818,7 @@ public class TableView implements TableOrView, Closeable {
         this.context.executeDelayedDisposal();
         long[] indexes = new long[columnIndexes.size()];
         for (int i = 0; i < columnIndexes.size(); i++) {
-            indexes[i] = columnIndexes.get(i).longValue();
+            indexes[i] = columnIndexes.get(i);
         }
         nativeDistinctMulti(nativePtr, indexes);
     }
